create utilities module
