+++ src/my_wordle_solver/entropy_soln.py
@@ -1,40 +1,11 @@
-import os
 import numpy as np
-import itertools as it
 from scipy import stats
 import json
+from utilities.common import WORD_LIST, PATTERN_COMBINATIONS
+import utilities.common
 
-#WORD_LIST = os.path.join(DATA_DIR, "possible_words.txt")
-#WORD_FREQ_FILE = os.path.join(DATA_DIR, "short_freqs.txt")
-
-d = {0: "NO", 1: "IN-WORD", 2: "EXACT"}
-
-WORD_LIST = []
-
-with open('possible_words.txt') as f:
-    for word in f.readlines():
-        WORD_LIST += [word.strip()]
-
-WORD_FREQ = dict()
-with open("short_freqs.txt") as fp:
-    for line in fp.readlines():
-        pieces = line.split(' ')
-        word = pieces[0]
-        freqs = [
-            float(piece.strip())
-            for piece in pieces[1:]
-        ]
-        WORD_FREQ[word] = np.mean(freqs[-5:])
-
-def find(s, ch):
-    return [i for i, ltr in enumerate(s) if ltr == ch]
-
-arr = [0, 1, 2]
-pattern_combinations = np.array(list(it.product(arr, repeat=5)))
-
-test_word_list = ["steal", "crepe", "erase", "abide", "irate"]
-test_comb = np.array([0,1,2,0,1])
-test_comb_d = [d[pattern] for pattern in test_comb]
+# NUMBERS TO PATTERN TRANSLATION
+# d = {0: "NO", 1: "IN-WORD", 2: "EXACT"}
 
 def patternmatch(wordcheck, test_word, pattern):
     equalmat = np.zeros(5)
@@ -56,7 +27,7 @@ def patternmatch(wordcheck, test_word, pattern):
             if checkmat[ind] ==0:
                 checkmat[ind] =1
                 equalmat[item] = 1
-            elif len(find(wordcheck, test_word[item]))>1:
+            elif len(utilities.common.find(wordcheck, test_word[item]))>1:
                 checkmat[wordcheck.index(test_word[item], ind +1)] =1
                 equalmat[item] = 1
             else:
@@ -84,20 +55,18 @@ def wordchecker(wordlist, test_word, pattern):
         output_d[wordlist[i]] = patternmatch(wordlist[i], test_word, pattern)
     return output_d
 
-#print(wordchecker(test_word_list, "crane", np.array([1,1,1,1,1])))
+
 
 def find_prob_distribution(wordlist, word):
     px = []
     tot = float(len(wordlist))
-    for pattern in pattern_combinations:
+    for pattern in PATTERN_COMBINATIONS:
         wordcheck_d = wordchecker(wordlist, word, pattern)
         num_true = sum(wordcheck_d.values())
         px += [num_true/tot]
         #print(tot)
     return np.array(px)
-prob_dist = find_prob_distribution(WORD_LIST, "stare")
-print(prob_dist)
-print(stats.entropy(prob_dist))
+
 
 def create_entropy_data(possible_words):
     ent = {}
@@ -111,3 +80,13 @@ def create_entropy_data(possible_words):
     return ent
 
 #ent_1 = create_entropy_data(WORD_LIST)
+
+# test_word_list = ["steal", "crepe", "erase", "abide", "irate"]
+# test_comb = np.array([0,1,2,0,1])
+# test_comb_d = [d[pattern] for pattern in test_comb]
+
+#print(wordchecker(test_word_list, "crane", np.array([1,1,1,1,1])))
+
+prob_dist = find_prob_distribution(WORD_LIST, "stare")
+print(prob_dist)
+print(stats.entropy(prob_dist))
