+++ src/utilities/common.py
@@ -1,0 +1,2 @@
+def find(s, ch):
+    return [i for i, ltr in enumerate(s) if ltr == ch]
